feat: Add namespace creation to Helm chart
This commit adds a Kubernetes Namespace resource definition to the Helm chart,
allowing the 'node-api' namespace to be created automatically if it does not
already exist when the chart is deployed.

This change supports GitOps workflows by ensuring the required namespace
is provisioned as part of the Helm chart deployment.